Improved functionality.

+++ src/WordTable.js
@@ -1,8 +1,10 @@
 import React, { useState } from 'react';
-import { TextField, Button, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, IconButton, Box, Dialog, DialogTitle, DialogContent, DialogActions, Typography } from '@mui/material';
+import DOMPurify from 'dompurify';
+import { TextField, Button, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, IconButton, Box, Dialog, DialogTitle, DialogContent, DialogActions, Typography, MenuItem, Select, FormControl, InputLabel } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/Delete';
 import RefreshIcon from '@mui/icons-material/Refresh';
-import VisibilityIcon from '@mui/icons-material/Visibility';
+import EditIcon from '@mui/icons-material/Edit'; // Edit button for editing words
+import { ArrowDropUp, ArrowDropDown } from '@mui/icons-material';
 import { getAllEntriesByLanguage, deleteEntry, updateEntry, saveEntry } from './dataStorage';
 
 function WordTable({ setMode }) {
@@ -11,6 +13,13 @@ function WordTable({ setMode }) {
   const [wordTable, setWordTable] = useState([]);
   const [selectedEntry, setSelectedEntry] = useState(null);
   const [sortConfig, setSortConfig] = useState({ key: null, direction: 'ascending' });
+  const [importFileName, setImportFileName] = useState(''); // For import file name
+  const [editMode, setEditMode] = useState(false); // To track if editing
+  const [editedCategory, setEditedCategory] = useState(''); // For editing category
+  const [editedLongTranslation, setEditedLongTranslation] = useState(''); // For editing long translation
+
+  const categories = ['None', 'Verbs', 'Adjectives', 'Family', 'Sport', 'Food', 'Travel', 'Work', 'Numbers', 'Colors']; // Example categories
+  const languages = ['Spanish', 'French', 'German']; // Dropdown options for languages
 
   const handleWordTableLoad = () => {
     const entries = selectedLanguage 
@@ -25,8 +34,10 @@ function WordTable({ setMode }) {
   };
 
   const handleDeleteEntry = (entry) => {
-    deleteEntry(entry);
-    handleWordTableLoad();
+    if (window.confirm(`Are you sure you want to delete the word "${entry.englishWord}"?`)) {
+      deleteEntry(entry);
+      handleWordTableLoad();
+    }
   };
 
   const handleResetScore = (entry) => {
@@ -38,12 +49,27 @@ function WordTable({ setMode }) {
     handleWordTableLoad();
   };
 
-  const handleViewDetails = (entry) => {
+  const handleEditEntry = (entry) => {
     setSelectedEntry(entry);
+    setEditedCategory(entry.category || 'None');
+    setEditedLongTranslation(entry.longTranslation);
+    setEditMode(true);
   };
 
   const handleCloseDetails = () => {
     setSelectedEntry(null);
+    setEditMode(false);
+  };
+
+  const handleSaveEdits = () => {
+    const updatedEntry = {
+      ...selectedEntry,
+      category: editedCategory,
+      longTranslation: DOMPurify.sanitize(editedLongTranslation),
+    };
+    updateEntry(updatedEntry);
+    handleWordTableLoad();
+    handleCloseDetails();
   };
 
   // Sorting functionality
@@ -80,6 +106,7 @@ function WordTable({ setMode }) {
   // Import words from a JSON file
   const handleImport = (event) => {
     const file = event.target.files[0];
+    setImportFileName(file.name); // Show file name
     const reader = new FileReader();
 
     reader.onload = (e) => {
@@ -95,24 +122,37 @@ function WordTable({ setMode }) {
 
   return (
     <Box>
-      <TextField
-        fullWidth
-        margin="normal"
-        label="Select Language"
-        variant="outlined"
-        value={selectedLanguage}
-        onChange={(e) => setSelectedLanguage(e.target.value)}
-        className="text-field"
-      />
-      <TextField
-        fullWidth
-        margin="normal"
-        label="Select Category"
-        variant="outlined"
-        value={selectedCategory}
-        onChange={(e) => setSelectedCategory(e.target.value)}
-        className="text-field"
-      />
+      {/* Dropdown for selecting language */}
+      <FormControl fullWidth margin="normal">
+        <InputLabel>Select Language</InputLabel>
+        <Select
+          value={selectedLanguage}
+          onChange={(e) => setSelectedLanguage(e.target.value)}
+          label="Select Language"
+        >
+          {languages.map((language) => (
+            <MenuItem key={language} value={language}>
+              {language}
+            </MenuItem>
+          ))}
+        </Select>
+      </FormControl>
+
+      {/* Dropdown for selecting category */}
+      <FormControl fullWidth margin="normal">
+        <InputLabel>Category</InputLabel>
+        <Select
+          value={selectedCategory}
+          onChange={(e) => setSelectedCategory(e.target.value)}
+        >
+          {categories.map((category) => (
+            <MenuItem key={category} value={category}>
+              {category}
+            </MenuItem>
+          ))}
+        </Select>
+      </FormControl>
+
       <Button
         variant="contained"
         fullWidth
@@ -147,16 +187,39 @@ function WordTable({ setMode }) {
         Import Words from File
         <input type="file" hidden onChange={handleImport} />
       </Button>
+      {importFileName && <Typography variant="body2" sx={{ marginTop: 1 }}>Selected file: {importFileName}</Typography>}
 
       <TableContainer sx={{ marginTop: 4 }}>
         <Table>
           <TableHead>
             <TableRow>
-              <TableCell className="table-header" onClick={() => handleSort('englishWord')}>English Word</TableCell>
-              <TableCell className="table-header" onClick={() => handleSort('singleTranslation')}>Translation</TableCell>
-              <TableCell className="table-header" onClick={() => handleSort('category')}>Category</TableCell>
-              <TableCell className="table-header" onClick={() => handleSort('points')}>Score</TableCell>
-              <TableCell className="table-header" onClick={() => handleSort('lastTested')}>Last Practiced</TableCell>
+              <TableCell className="table-header" onClick={() => handleSort('englishWord')}>
+                English Word
+                {sortConfig.key === 'englishWord' && (sortConfig.direction === 'ascending' ? <ArrowDropUp /> : <ArrowDropDown />)}
+              </TableCell>
+              <TableCell className="table-header" onClick={() => handleSort('singleTranslation')}>
+                Translation
+                {sortConfig.key === 'singleTranslation' && (sortConfig.direction === 'ascending' ? <ArrowDropUp /> : <ArrowDropDown />)}
+              </TableCell>
+              <TableCell className="table-header" onClick={() => handleSort('category')}>
+                Category
+                {sortConfig.key === 'category' && (sortConfig.direction === 'ascending' ? <ArrowDropUp /> : <ArrowDropDown />)}
+              </TableCell>
+              {/* Conditionally render the Language column if no language is selected */}
+              {!selectedLanguage && (
+                <TableCell className="table-header" onClick={() => handleSort('learningLanguage')}>
+                  Language
+                  {sortConfig.key === 'learningLanguage' && (sortConfig.direction === 'ascending' ? <ArrowDropUp /> : <ArrowDropDown />)}
+                </TableCell>
+              )}
+              <TableCell className="table-header" onClick={() => handleSort('points')}>
+                Score
+                {sortConfig.key === 'points' && (sortConfig.direction === 'ascending' ? <ArrowDropUp /> : <ArrowDropDown />)}
+              </TableCell>
+              <TableCell className="table-header" onClick={() => handleSort('lastTested')}>
+                Last Practiced
+                {sortConfig.key === 'lastTested' && (sortConfig.direction === 'ascending' ? <ArrowDropUp /> : <ArrowDropDown />)}
+              </TableCell>
               <TableCell className="table-header">Actions</TableCell>
             </TableRow>
           </TableHead>
@@ -166,11 +229,13 @@ function WordTable({ setMode }) {
                 <TableCell>{entry.englishWord}</TableCell>
                 <TableCell>{entry.singleTranslation}</TableCell>
                 <TableCell>{entry.category || 'None'}</TableCell>
+                {/* Conditionally render the Language cell if no language is selected */}
+                {!selectedLanguage && <TableCell>{entry.learningLanguage}</TableCell>}
                 <TableCell>{entry.points}</TableCell>
                 <TableCell>{entry.lastTested ? new Date(entry.lastTested).toLocaleDateString() : 'Never'}</TableCell>
                 <TableCell>
-                  <IconButton onClick={() => handleViewDetails(entry)} color="primary">
-                    <VisibilityIcon />
+                  <IconButton onClick={() => handleEditEntry(entry)} color="primary">
+                    <EditIcon />
                   </IconButton>
                   <IconButton onClick={() => handleDeleteEntry(entry)} color="error">
                     <DeleteIcon />
@@ -193,7 +258,7 @@ function WordTable({ setMode }) {
         onClick={() => setMode('input')}
         className="button"
       >
-        Switch to Input Mode
+        Add to Dictionary
       </Button>
       <Button
         variant="outlined"
@@ -203,32 +268,45 @@ function WordTable({ setMode }) {
         onClick={() => setMode('learning')}
         className="button"
       >
-        Switch to Learning Mode
+        Start Learning
       </Button>
 
-      {/* Details Dialog */}
-      <Dialog open={Boolean(selectedEntry)} onClose={handleCloseDetails}>
-        <DialogTitle>Translation Details</DialogTitle>
-        <DialogContent>
-          {selectedEntry && (
-            <Box>
-              <Typography variant="h6">English Word:</Typography>
-              <Typography>{selectedEntry.englishWord}</Typography>
-              <Typography variant="h6" sx={{ marginTop: 2 }}>Single Translation:</Typography>
-              <Typography>{selectedEntry.singleTranslation}</Typography>
-              <Typography variant="h6" sx={{ marginTop: 2 }}>Category:</Typography>
-              <Typography>{selectedEntry.category || 'None'}</Typography>
-              <Typography variant="h6" sx={{ marginTop: 2 }}>Longer Translation and Notes:</Typography>
-              {selectedEntry.longTranslation.split('\n').map((line, index) => (
-                <Typography key={index} variant="body1">{line}</Typography>
-              ))}
-            </Box>
-          )}
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleCloseDetails} color="primary">Close</Button>
-        </DialogActions>
-      </Dialog>
+      {/* Edit Dialog */}
+      {editMode && selectedEntry && (
+        <Dialog open={editMode} onClose={handleCloseDetails}>
+          <DialogTitle>Edit Entry</DialogTitle>
+          <DialogContent>
+            <Typography variant="h6">English Word: {selectedEntry.englishWord}</Typography>
+            <FormControl fullWidth margin="normal">
+              <InputLabel>Category</InputLabel>
+              <Select
+                value={editedCategory}
+                onChange={(e) => setEditedCategory(e.target.value)}
+              >
+                {categories.map((category) => (
+                  <MenuItem key={category} value={category}>
+                    {category}
+                  </MenuItem>
+                ))}
+              </Select>
+            </FormControl>
+            <TextField
+              fullWidth
+              margin="normal"
+              label="Longer Translation and Notes"
+              multiline
+              rows={4}
+              variant="outlined"
+              value={editedLongTranslation}
+              onChange={(e) => setEditedLongTranslation(e.target.value)}
+            />
+          </DialogContent>
+          <DialogActions>
+            <Button onClick={handleCloseDetails} color="secondary">Cancel</Button>
+            <Button onClick={handleSaveEdits} color="primary">Save Changes</Button>
+          </DialogActions>
+        </Dialog>
+      )}
     </Box>
   );
 }
